🛠 Fix overflowing integers

+++ Sources/WebFoundation/Blob.swift
@@ -75,7 +75,7 @@ public class Blob: JSClass, CustomStringConvertible {
     ///
     /// - Returns: A new `Blob` object containing the specified subset of the data
     /// contained within the blob on which this method was called. The original blob is not altered.
-    public func slice(start: Int? = nil, end: Int? = nil, contentType: String? = nil) -> Blob? {
+    public func slice(start: Int64? = nil, end: Int64? = nil, contentType: String? = nil) -> Blob? {
         guard let newBlob = jsValue.slice.function?.callAsFunction(this: jsValue.object, start, end, contentType) else { return nil }
         return .init(newBlob)
     }
